refactor: restructure directory layout for narrative clarity
5 top-level modules tell the system's story at a glance:
- agent/     → definition, session, capabilities (tools, skills)
- backends/  → AI backend adapters (+ model-maps)
- workflow/  → multi-agent orchestration (+ context)
- daemon/    → background service (+ server from cli/)
- cli/       → command-line interface

Key moves:
- core/ → agent/ (tools/ and skills/ nested under it)
- model-maps.ts → backends/ (where it belongs)
- context/ → workflow/context/ (it's a workflow concept)
- cli/server.ts → daemon/server.ts (it's daemon infrastructure)
- workflow/mcp-config.ts → workflow/backend-config.ts
