+++ packages/agent-worker/src/agent/skills/provider.ts
