+++ packages/agent-worker/src/workflow/controller/types.ts
@@ -4,8 +4,8 @@
  */
 
 import type { ResolvedAgent } from '../types.ts'
-import type { ContextProvider } from '../../context/provider.ts'
-import type { Message, InboxMessage } from '../../context/types.ts'
+import type { ContextProvider } from '../context/provider.ts'
+import type { Message, InboxMessage } from '../context/types.ts'
 import type { Backend } from '../../backends/types.ts'
 
 // ==================== Controller ====================
