+++ packages/agent-worker/src/workflow/index.ts
@@ -6,6 +6,6 @@ export * from './types.ts'
 export * from './parser.ts'
 export * from './interpolate.ts'
 export * from './runner.ts'
-export * from './mcp-config.ts'
+export * from './backend-config.ts'
 export * from './controller/index.ts'
 export * from './logger.ts'
