+++ packages/agent-worker/src/workflow/types.ts
@@ -2,7 +2,7 @@
  * Workflow file type definitions
  */
 
-import type { ContextConfig } from '../context/types.js'
+import type { ContextConfig } from './context/types.js'
 
 // Re-export context types for convenience
 export type {
