+++ packages/agent-worker/test/mock-backend.test.ts
@@ -9,9 +9,9 @@
 
 import { describe, test, expect } from 'bun:test'
 import { validateWorkflow } from '../src/workflow/parser.ts'
-import { createMemoryContextProvider } from '../src/context/memory-provider.ts'
-import { createContextMCPServer } from '../src/context/mcp-server.ts'
-import { runWithHttp } from '../src/context/http-transport.ts'
+import { createMemoryContextProvider } from '../src/workflow/context/memory-provider.ts'
+import { createContextMCPServer } from '../src/workflow/context/mcp-server.ts'
+import { runWithHttp } from '../src/workflow/context/http-transport.ts'
 import { runMockAgent } from '../src/workflow/controller/mock-runner.ts'
 import type { AgentRunContext } from '../src/workflow/controller/types.ts'
